Update the preferredContentSize of the RichTextViewController whenever the UITextView's contentSize changes
Previously, the preferredContentSize calculation was broken: the layout and render process is not taking place synchroneously
so that the process might not have completed when the dispatched block inside renderDocumentIfNeeded() invokes the
calculateAndSetPreferredContentSize() function.

Whenever the RichTextViewController is initialized with isScrollEnabled: true (or when no value is passed), we can monitor
the contentSize of the UITextView for changes and update the preferredContentSize accordingly.